refactor(commands): extract maintenance commands into maintenance.rs
